move BtsAdapter
