+++ AIMSICD/src/main/java/com/secupwn/aimsicd/ui/fragments/DbViewerFragment.java
@@ -11,9 +11,9 @@ import android.widget.Spinner;
 import com.secupwn.aimsicd.R;
 import com.secupwn.aimsicd.utils.RealmHelper;
 import com.secupwn.aimsicd.adapters.DbViewerSpinnerAdapter;
-import com.secupwn.aimsicd.adapters.MeasureAdapter;
+import com.secupwn.aimsicd.data.adapter.MeasureAdapter;
 import com.secupwn.aimsicd.adapters.MeasuredCellStrengthAdapter;
-import com.secupwn.aimsicd.adapters.UniqueBtsAdapter;
+import com.secupwn.aimsicd.data.adapter.BtsAdapter;
 import com.secupwn.aimsicd.data.adapter.DefaultLocationAdapter;
 import com.secupwn.aimsicd.data.adapter.DetectionStringAdapter;
 import com.secupwn.aimsicd.data.adapter.EventAdapter;
@@ -83,7 +83,7 @@ public final class DbViewerFragment extends InjectionFragment {
 
                 switch (position) {
                     case 0:
-                        setListAdapter(new UniqueBtsAdapter(getActivity(), realm.allObjects(BTS.class), true));
+                        setListAdapter(new BtsAdapter(getActivity(), realm.allObjects(BTS.class), true));
                         break;
                     case 1:
                         setListAdapter(new MeasureAdapter(getActivity(), realm.allObjects(Measure.class), true));
